refactor: grpc code structure
