+++ crates/geo-ip/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "@ec-url-store/geo-ip",
+  "version": "0.0.0",
+  "scripts": {
+    "build": "cargo build",
+    "build:release": "cargo build --release",
+    "dev": "cargo run",
+    "start": "cargo run --release"
+  }
+}
