+++ crates/geo-ip/src/main.rs
@@ -1,9 +1,5 @@
 mod libs;
-mod gen;
 mod utils;
-mod gen_dep {
-    pub(crate) const FILE_DESCRIPTOR_SET: &[u8] = include_bytes!("./gen/proto_descriptor.bin");
-}
 
 #[tokio::main]
 async fn main() {
