+++ crates/grpc_proto/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "geo-ip"
+name = "grpc_proto"
 version = "0.1.0"
 edition = "2021"
 
@@ -14,7 +14,6 @@ async-stream = { version = "0.3", optional = true }
 tokio-util = { version = "0.7.8", optional = true }
 maxminddb = "0.24.0"
 dotenv = "0.15.0"
-grpc_proto = { path = "../grpc_proto" }
 
 [build-dependencies]
 tonic-build = "0.12.3"
