+++ crates/grpc_proto/src/gen/proto_descriptor.bin
